도시, 도로 null 검증 추가

+++ ktor-app/src/main/kotlin/com/study/weather/model/WeatherRequest.kt
@@ -29,6 +29,6 @@ data class WeatherRequest(
     val dataType: String = "json",
     var baseDate:String = "",
     var basetime:String = "",
-    var nx:Double = 0.0,
-    var ny:Double = 0.0
+    var nx:Int = 0,
+    var ny:Int = 0
     )
